re-refactor

+++ credit/preblock/concat.py
@@ -138,10 +138,10 @@ class ConcatToTensor(BasePreblock):
         accel = next((t.device for t in input_tensors if t.device.type != "cpu"), None)
         if accel is not None:
             input_tensors = [t.to(accel) for t in input_tensors]
-        input_tensor = torch.cat(input_tensors, dim=1)
+        input_tensor = torch.cat(input_tensors, dim=1).float()
 
         if target_tensors:
-            target_tensor = torch.cat(target_tensors, dim=1)
+            target_tensor = torch.cat(target_tensors, dim=1).float()
             return input_tensor, target_tensor, metadata
 
         return input_tensor, metadata
